+++ dart/geobase/CHANGELOG.md
@@ -14,12 +14,14 @@ NOTE: Version 0.6.0 currently [under development](https://github.com/navibyte/ge
 * [Simplify Feature and FeatureCollection types #195](https://github.com/navibyte/geospatial/issues/195)
 * [Remove deprecated types in geobase version 0.6.0 #194](https://github.com/navibyte/geospatial/issues/194)
 * [Consistent crs and trs references in meta classes #196](https://github.com/navibyte/geospatial/issues/196)
+* [Deprecate PositionArray, PositionCoords and BoxCoords #201](https://github.com/navibyte/geospatial/issues/201)
 
 🛠 Refactoring:
 * isEmptyByGeometry on Bounded (and feature objects too, not only geometries)
 
 🛠 Maintenance:
-* PositionArray, PositionCoords and BoxCoords moved from vector_data to vector.
+* PositionArray, PositionCoords and BoxCoords moved from vector_data to vector
+  (and later deprecated by #201).
 
 ## 0.5.1